move constants, make audiogram_chart not use bloc
